+++ credcollect/templates/index.html
@@ -1,91 +1,174 @@
-<html>
-<style>
-body { padding: 2em; }
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <title>Madcore</title>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+  <style>
+    /* Remove the navbar's default margin-bottom and rounded borders */ 
+    .navbar {
+      margin-bottom: 0;
+      border-radius: 0;
+    }
+    
+    /* Set height of the grid so .sidenav can be 100% (adjust as needed) */
+    .row.content {height: 450px}
+    
+    /* Set gray background color and 100% height */
+    .sidenav {
+      padding-top: 20px;
+      background-color: #f1f1f1;
+      height: 100%;
+    }
+    
+    /* Set black background color, white text and some padding */
+    footer {
+      background-color: #555;
+      color: white;
+      padding: 15px;
+    }
+    
+    /* On small screens, set height to 'auto' for sidenav and grid */
+    @media screen and (max-width: 767px) {
+      .sidenav {
+        height: auto;
+        padding: 15px;
+      }
+      .row.content {height:auto;} 
+    }
+    .loginBtn {
+      box-sizing: border-box;
+      position: relative;
+      /* width: 13em;  - apply for fixed size */
+      margin: 0.2em;
+      padding: 0 15px 0 46px;
+      border: none;
+      text-align: left;
+      line-height: 34px;
+      white-space: nowrap;
+      border-radius: 0.2em;
+      font-size: 16px;
+      color: #FFF;
+    }
+    .loginBtn:before {
+      content: "";
+      box-sizing: border-box;
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 34px;
+      height: 100%;
+    }
+    .loginBtn:focus {
+      outline: none;
+    }
+    .loginBtn:active {
+      box-shadow: inset 0 0 0 32px rgba(0,0,0,0.1);
+    }
+
+    /* Google */
+    .loginBtn--google {
+      /*font-family: "Roboto", Roboto, arial, sans-serif;*/
+      background: #DD4B39;
+    }
+    .loginBtn--google:before {
+      border-right: #BB3F30 1px solid;
+      background: url('https://s3-us-west-2.amazonaws.com/s.cdpn.io/14082/icon_google.png') 6px 6px no-repeat;
+    }
+    .loginBtn--google:hover,
+    .loginBtn--google:focus {
+      background: #E74B37;
+    }
+
+    /* Github */
+    .loginBtn--git {
+      /*font-family: "Roboto", Roboto, arial, sans-serif;*/
+      background: #2bd640;
+    }
+    .loginBtn--git:before {
+      border-right:#2bd640 1px solid;
+      background: url('https://static.wixstatic.com/media/d333e2_57d24de2ed6f434198cc9a7d96a961c7~mv2.png');
+      background-size: cover;
+    }
+    .loginBtn--git:hover,
+    .loginBtn--git:focus {
+      background: #2bd640;
+    }
+  </style>
+</head>
+<body>
+
+<nav class="navbar navbar-inverse">
+  <div class="container-fluid">
+    <div class="navbar-header">
+      <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#myNavbar">
+        <span class="icon-bar"></span>
+        <span class="icon-bar"></span>
+        <span class="icon-bar"></span>                        
+      </button>
+      <a class="navbar-brand" href="#"> <img src="https://madcore.ai/wp-content/uploads/2016/12/MadCoreLogo001-e1481323055628.png" style="max-width:100px; margin-top: -7px;background: #000;"></a>
+    </div>
+    <div class="collapse navbar-collapse" id="myNavbar">
+      <!-- <ul class="nav navbar-nav">
+        <li class="active"><a href="#">Home</a></li>
+        <li><a href="#">About</a></li>
+        <li><a href="#">Projects</a></li>
+        <li><a href="#">Contact</a></li>
+      </ul> -->
+      {% if session.google_token %}
+      <ul class="nav navbar-nav navbar-right">
+        <li><a href="/logout"><span class="glyphicon glyphicon-log-in"></span> Logout</a></li>
+      </ul>
+      {% endif %}
+    </div>
+  </div>
+</nav>
+  
+<div class="container-fluid text-center">    
+  <div class="row content">
+    <div class="col-sm-12"> 
+      <h1>Welcome</h1>
+      <hr>
+      <h3>Social Login</h3>
+      {% if google_status %}
+        <div>
+          <button class="loginBtn loginBtn--google" onclick="location.href='/login/google';">
+            Login with Google
+          </button>
+        </div>
+      {% endif %}
+      {% if git_status %}
+        <div >
+          <button class="loginBtn loginBtn--git" onclick="location.href='/login/github';">
+            Login with Github
+          </button>
+        </div>
+      {% endif %}
+      <div>
+        {% if message %}
+        <h3>{{message}}</h3>
+        <a href="/logout"><h3>Logout<h3></a>
+        {% endif %}
+    </div>
+  </div>
+</div>
+
+<footer class="container-fluid text-center">
+  <p>Madcore</p>
+</footer>
+
+</body>
+</html>
 
 
-/* Shared */
-.loginBtn {
-  box-sizing: border-box;
-  position: relative;
-  /* width: 13em;  - apply for fixed size */
-  margin: 0.2em;
-  padding: 0 15px 0 46px;
-  border: none;
-  text-align: left;
-  line-height: 34px;
-  white-space: nowrap;
-  border-radius: 0.2em;
-  font-size: 16px;
-  color: #FFF;
-}
-.loginBtn:before {
-  content: "";
-  box-sizing: border-box;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 34px;
-  height: 100%;
-}
-.loginBtn:focus {
-  outline: none;
-}
-.loginBtn:active {
-  box-shadow: inset 0 0 0 32px rgba(0,0,0,0.1);
-}
-
-/* Google */
-.loginBtn--google {
-  /*font-family: "Roboto", Roboto, arial, sans-serif;*/
-  background: #DD4B39;
-}
-.loginBtn--google:before {
-  border-right: #BB3F30 1px solid;
-  background: url('https://s3-us-west-2.amazonaws.com/s.cdpn.io/14082/icon_google.png') 6px 6px no-repeat;
-}
-.loginBtn--google:hover,
-.loginBtn--google:focus {
-  background: #E74B37;
-}
-
-/* Github */
-.loginBtn--git {
-  /*font-family: "Roboto", Roboto, arial, sans-serif;*/
-  background: #2bd640;
-}
-.loginBtn--git:before {
-  border-right:#2bd640 1px solid;
-  background: url('https://static.wixstatic.com/media/d333e2_57d24de2ed6f434198cc9a7d96a961c7~mv2.png');
-  background-size: cover;
-}
-.loginBtn--git:hover,
-.loginBtn--git:focus {
-  background: #2bd640;
-}
-</style>
 <body>
 <div>
-  <img src="https://madcore.ai/wp-content/uploads/2016/12/MadCoreLogo001-e1481323055628.png" style="background: #000">
+ 
 </div>
-{% if google_status %}
-  <div style="float: left">
-    <button class="loginBtn loginBtn--google" onclick="location.href='/login/google';">
-      Login with Google
-    </button>
-  </div>
-{% endif %}
-{% if git_status %}
-  <div style="float: left">
-    <button class="loginBtn loginBtn--git" onclick="location.href='/login/github';">
-      Login with Github
-    </button>
-  </div>
-{% endif %}
-<div>
-  <h3>{{message}}</h3>
-  {% if message %}
-  <a href="/logout"><h3>Logout<h3></a>
-  {% endif %}
+
 </div>
 </body>
 </html>
